Move Glide to its own module
