+++ ui/glide/README.md
@@ -1,0 +1,3 @@
+# :ui:glide
+
+Configuration and ModelLoaders for the Glide image loading library.
