+++ ui/glide/build.gradle
@@ -1,0 +1,27 @@
+plugins {
+    id("com.android.library")
+}
+apply from: "../../common.gradle"
+apply from: "../../playFlavor.gradle"
+
+android {
+    lintOptions {
+        disable "InvalidPeriodicWorkRequestInterval", "ObsoleteLintCustomCheck", "DefaultLocale", "UnusedAttribute",
+                "ParcelClassLoader", "CheckResult", "TrustAllX509TrustManager",
+                "StaticFieldLeak", "IconDensities", "IconDuplicates", "MissingPermission", "AppCompatResource",
+                "GradleCompatible", "QueryPermissionsNeeded"
+    }
+}
+
+dependencies {
+    implementation project(":model")
+    implementation project(":core")
+
+    implementation "androidx.palette:palette:$paletteVersion"
+
+    implementation "com.github.bumptech.glide:glide:$glideVersion"
+    implementation "com.github.bumptech.glide:okhttp3-integration:$glideVersion@aar"
+    annotationProcessor "com.github.bumptech.glide:compiler:$glideVersion"
+    implementation "com.squareup.okhttp3:okhttp:$okhttpVersion"
+    implementation "commons-io:commons-io:$commonsioVersion"
+}
