+++ ui/glide/src/main/AndroidManifest.xml
@@ -1,0 +1,1 @@
+<manifest package="de.danoeh.antennapod.ui.glide" />
