+++ ui/glide/src/main/java/de/danoeh/antennapod/ui/glide/MetadataRetrieverLoader.java
@@ -1,4 +1,4 @@
-package de.danoeh.antennapod.core.glide;
+package de.danoeh.antennapod.ui.glide;
 
 import android.content.Context;
 import androidx.annotation.NonNull;
