+++ ui/glide/src/main/java/de/danoeh/antennapod/ui/glide/PaletteBitmapResource.java
@@ -2,7 +2,7 @@
  * Source: https://github.com/bumptech/glide/wiki/Custom-targets#palette-example
  */
 
-package de.danoeh.antennapod.core.glide;
+package de.danoeh.antennapod.ui.glide;
 
 import androidx.annotation.NonNull;
