+++ ui/statistics/build.gradle
@@ -28,7 +28,6 @@ dependencies {
 
     implementation "org.greenrobot:eventbus:$eventbusVersion"
     implementation "com.github.bumptech.glide:glide:$glideVersion"
-    annotationProcessor "com.github.bumptech.glide:compiler:$glideVersion"
     implementation "io.reactivex.rxjava2:rxandroid:$rxAndroidVersion"
     implementation "io.reactivex.rxjava2:rxjava:$rxJavaVersion"
     implementation "com.joanzapata.iconify:android-iconify-fontawesome:$iconifyVersion"
